refactoring: make utils robust and easier to read
